pyproject.toml
[tool.poetry]
name = "multi-llm"
version = "0.1.0"
description = ""
authors = ["Alonso Isidoro Román <alonsoir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyarrow = "^17.0.0"
streamlit = "^1.36.0"
python-dotenv = "^1.0.1"
openai = "^1.37.0"
mem0ai = "^0.0.9"
litellm = "^1.41.25"
pymilvus = "^2.4.4"
scikit-learn = "^1.5.1"
pinecone-client = "^5.0.0"
transformers = "^4.43.1"
psutil = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
